Fix factory bot implementation
- Change the type of total column

+++ db/migrate/20220602162050_create_reservations.rb
@@ -1,7 +1,7 @@
 class CreateReservations < ActiveRecord::Migration[7.0]
   def change
     create_table :reservations do |t|
-      t.float :total
+      t.decimal :total
       t.datetime :reserved_date
       t.string :city
       t.integer :user_id
